+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.1.2](https://gitlab.com/redsky_public/framework/compare/v1.1.1...v1.1.2) (2022-12-24)
+
+### Bug Fixes
+
+-   **box:** Added other missing events ([1e1d455](https://gitlab.com/redsky_public/framework/commit/1e1d455eaf49ba6804e9f165a041c807694b8feb))
+
+### [1.1.1](https://gitlab.com/redsky_public/framework/compare/v1.1.0...v1.1.1) (2022-12-24)
+
+### Bug Fixes
+
+-   **framework-settings:** Made overrides optional ([c361ae8](https://gitlab.com/redsky_public/framework/commit/c361ae8a54689582d1ee63fd25da128bf60b0362))
+
+## [1.1.0](https://gitlab.com/redsky_public/framework/compare/v1.0.0...v1.1.0) (2022-12-24)
+
+### Features
+
+-   **box:** Improved auto-complete, added other mouse, touch, pointer events ([bf29edf](https://gitlab.com/redsky_public/framework/commit/bf29edfed6b2e23be5b2614c4f1a51d5947522c8))
+
 ## [1.0.0](https://gitlab.com/redsky_public/framework/compare/v0.16.1...v1.0.0) (2022-12-21)
 
 ### Features